pyproject.toml
[tool.poetry]
name = "parking-lot-api"
version = "0.1.0"
description = ""
authors = ["Kevin Jennison <kevin.jennison1@gmail.com>"]
readme = "README.md"
packages = [{include = "parking_lot_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-filter = "^23.1"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
